% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eprime_find.R
\name{eprime_find}
\alias{eprime_find}
\title{Find eprime files}
\usage{
eprime_find(in_dir = "~/LCBC/Projects/Cross_projects/computer_tasks",
  data_dir = "C_Eprime_Testrom1",
  out_dir = "~/LCBC/Users/athanasm_mo/", task = "Antisaccade",
  copy_etxts = 1, copy_edats = 1)
}
\arguments{
\item{in_dir}{Directory with E-prime files}

\item{data_dir}{Sub-dir of base that has the data}

\item{out_dir}{directory to output the data to}

\item{task}{Which task, one of ["Antisaccade", "n-back", "Attention"]}

\item{copy_etxts}{binary logical, 1 to copy the txt's}

\item{copy_edats}{binary logical, 1 to copy the edat's.}
}
\description{
Find eprime files
}
