% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eprime_parse.R
\name{eprime_parse}
\alias{eprime_parse}
\title{Parse LCBC E-prime files}
\usage{

  eprime_parse(in_dir = "/Users/athanasm/LCBC/Projects/Cross_projects/computer_tasks",
  out_dir = "/Users/athanasm/Desktop/tasks", task = "n-back",
  MOAS = "/Users/athanasm/LCBC/Projects/Cross_projects/MOAS/Data/MOAS.RData",
  quietly = FALSE)
}
\arguments{
\item{in_dir}{Directory with E-prime files}

\item{out_dir}{directory to output the data to}

\item{task}{Which task, one of ["Antisaccade", "n-back", "Attention"]}

\item{MOAS}{Either string with path to, or as data.frame, the MOAS dataframe}

\item{quietly}{logical if log text should be printed in the console}
}
\description{
Will look for new eprime files, and start
logging information about them.
}
