% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eprime_parser.R
\name{eprime_parser}
\alias{eprime_parser}
\title{Parse LCBC E-prime files}
\usage{
eprime_parser(in_dir = "~/LCBC/Projects/Cross_projects/computer_tasks",
  out_dir = "~/LCBC/Users/athanasm_mo", data_dir = "C_Eprime_Testrom1",
  task = "Antisaccade",
  MOAS = "~/LCBC/Projects/Cross_projects/MOAS/Data/MOAS.RData")
}
\arguments{
\item{in_dir}{Directory with E-prime files}

\item{out_dir}{directory to output the data to}

\item{data_dir}{Sub-dir of base that has the data}

\item{task}{Which task, one of ["Antisaccade", "n-back", "Attention"]}

\item{MOAS}{Either string with path to, or as data.frame, the MOAS dataframe}
}
\description{
Parse LCBC E-prime files
}
