% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eprime_process.R
\name{eprime_process_all}
\alias{eprime_process_all}
\title{Process all files}
\usage{
eprime_process_all(ff_df, in_dir, task, MOAS, paths, logs,
  quietly = FALSE)
}
\arguments{
\item{ff_df}{data frame of files to process from status.log}

\item{in_dir}{Directory with E-prime files}

\item{task}{Which task, one of ["Antisaccade", "n-back", "Attention"]}

\item{MOAS}{Either string with path to, or as data.frame, the MOAS dataframe}

\item{paths}{list of paths from [\code{eprime_setup_dirs}]}

\item{logs}{list of paths from [\code{eprime_setup_logs}]}

\item{quietly}{logical if log text should be printed in the console}
}
\description{
Calls eprime_process_file and runs through each file listed in the
provided data.frame to process.
}
