% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eprime_process.R
\name{eprime_process_file}
\alias{eprime_process_file}
\title{Process an E-prime file for LCBC}
\usage{
eprime_process_file(efile_path, in_dir, task, MOAS, paths, logs,
  start_message = "Starting processing for", quietly = FALSE,
  return_options = FALSE)
}
\arguments{
\item{efile_path}{path to E-prime txt file}

\item{in_dir}{Directory with E-prime files}

\item{task}{Which task, one of ["Antisaccade", "n-back", "Attention"]}

\item{MOAS}{Either string with path to, or as data.frame, the MOAS dataframe}

\item{paths}{list of paths from [\code{eprime_setup_dirs}]}

\item{logs}{list of paths from [\code{eprime_setup_logs}]}

\item{start_message}{string message to add to log for this action}

\item{quietly}{logical if log text should be printed in the console}

\item{return_options}{logical if options generated for the file should
be returned to user}
}
\description{
Runs the processing of E-prime files for LCBC. 
This will check the existance of files, any
previous logs for duplicated entries etc.
}
