% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eprime_resolve.R
\name{eprime_resolve_all}
\alias{eprime_resolve_all}
\title{Attempt manual resolve on all files that are not completed}
\usage{
eprime_resolve_all(in_dir, out_dir, task, MOAS,
  gmail_user = "athanasiamo@gmail.com", .choice = NULL)
}
\arguments{
\item{in_dir}{Directory with E-prime files}

\item{out_dir}{directory to output the data to}

\item{task}{Which task, one of ["Antisaccade", "n-back", "Attention"]}

\item{MOAS}{Either string with path to, or as data.frame, the MOAS dataframe}

\item{gmail_user}{Email with access to the LCBC test calendar
on google}

\item{.choice}{never to be set by the user. Created for unit testing only}
}
\description{
Will check the status.log for files that need resolution,
and attempt to resolve their issues. This inclused attempting
to process the files automatically before prompting for manual 
input to resolve the issues that are stopping the file from
being processed.
}
