% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eprime-utils.R
\name{eprime_setup_dirs}
\alias{eprime_setup_dirs}
\title{Set up directories for E-prime parsing}
\usage{
eprime_setup_dirs(path)
}
\arguments{
\item{path}{Path to where the directories should be
placed}
}
\value{
list of paths
}
\description{
Sets up the necessary folder structure, if it
does not already exists, for the eprime
parser to work
}
