% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log-funcs.R
\name{eprime_setup_logs}
\alias{eprime_setup_logs}
\title{Set up logs for eprime parser}
\usage{
eprime_setup_logs(path)
}
\arguments{
\item{path}{Path to where the logs should be}
}
\value{
list of log paths
}
\description{
Initiates and creates paths to necessary
log-files for the E-prime parser
}
