% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eprime_setups.R
\name{eprime_setup_paths}
\alias{eprime_setup_paths}
\title{Set up directories for E-prime parsing}
\usage{
eprime_setup_paths(out_dir, task, logs, quietly)
}
\arguments{
\item{out_dir}{directory to output the data to}

\item{task}{Which task, one of ["Antisaccade", "n-back", "Attention"]}

\item{logs}{list of paths from [\code{eprime_setup_logs}]}

\item{quietly}{logical if log text should be printed in the console}
}
\value{
list of paths
}
\description{
Sets up the necessary folder structure, if it
does not already exists, for the eprime
parser to work
}
