% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eprime_find.R
\name{eprime_find}
\alias{eprime_find}
\title{Find eprime files}
\usage{
eprime_find(in_dir, task, paths, logs, quietly = FALSE)
}
\arguments{
\item{in_dir}{Directory with E-prime files}

\item{task}{Which task, one of ["Antisaccade", "n-back", "Attention"]}

\item{paths}{list of paths from [\code{eprime_setup_dirs}]}

\item{logs}{list of paths from [\code{eprime_setup_logs}]}

\item{quietly}{logical if log text should be printed in the console}
}
\description{
Find eprime files
}
