% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eprime_process_file.R
\name{eprime_process_file}
\alias{eprime_process_file}
\title{Process an E-prime file for LCBC}
\usage{
eprime_process_file(efile, MOAS, task, paths, logs, errorlines, loglines)
}
\arguments{
\item{efile}{path to E-prime txt file}

\item{MOAS}{Either string with path to, or as data.frame, the MOAS dataframe}

\item{task}{Which task, one of ["Antisaccade", "n-back", "Attention"]}

\item{paths}{list of paths from [\code{eprime_setup_dirs}]}

\item{logs}{list of paths from [\code{eprime_setup_logs}]}

\item{errorlines}{errorlines from [\code{eprime_setup_logs}]}

\item{loglines}{loglines from [\code{eprime_setup_logs}]}
}
\description{
Runs the processing of E-prime files for LCBC. 
This will check the existance of files, any
previous logs for duplicated entries etc.
}
